(fix) cast divide result into int
This problem is due to a behaviour difference between python2 and python3:
In python2, a division by int values returns a int as result (even if the result should be float).

In python3 the result is cast in float value

(cherry picked from commit a489178fa0bfec2c1cd111e486faf98ee56abc07)

+++ services/mmc/plugins/glpi/database_93.py
@@ -3023,7 +3023,7 @@ class Glpi93(DyngroupDatabaseHelper):
             thanks to http://stackoverflow.com/questions/4130922/how-to-increment-datetime-month-in-python
             """
             month = sourcedate.month - 1 + months
-            year = sourcedate.year + month / 12
+            year = int(sourcedate.year + month / 12)
             month = month % 12 + 1
             day = min(sourcedate.day, calendar.monthrange(year, month)[1])
             return datetime.date(year, month, day)
